libhfcommon/util: disable util_getProgAddr() for cygwin as it doesn't use ELF

+++ sanitizers.c
@@ -185,7 +185,7 @@ size_t sanitizers_parseReport(run_t* run, pid_t pid, funcs_t* funcs, uint64_t* p
         } else {
             char* pLineLC = lineptr;
             /* Trim leading spaces */
-            while (*pLineLC != '\0' && isspace(*pLineLC)) {
+            while (*pLineLC != '\0' && isspace((int)*pLineLC)) {
                 ++pLineLC;
             }
